Сделал картинки в слайдере новостей на главной кликабельными
