+++ src/components/NewsSlider/NewsSlider.module.scss
